response fix

+++ app/controllers/notificationsControllers.js
@@ -1,13 +1,22 @@
 const notifService = require('../services/notifService');
+const {v4: uuid} = require('uuid');
 
 module.exports = {
     getAllNotif(req, res) {
             notifService.findAll()
             .then(notif => {
-                return notif;
+                res.status(200).json({
+                    msg: "success get all notifications",
+                    status: 200,
+                    data: notif
+                });
             })
             .catch(err => {
-                return err;
+                res.status(500).json({
+                    msg: "error get all notifications",
+                    status: 500,
+                    err
+                });
             });
     },
     getNotifByUserId(req, res) {
@@ -29,24 +38,55 @@ module.exports = {
             });
         });
     },
-    deleteNotifById(req, res) {
+    async deleteNotifById(req, res) {
         const {id} = req.params
+        let check = await notifService.findByPk(id)
+        if(!check){
+            res.status(404).json({
+                msg: "data not found",
+                status: 404,
+            });
+            return
+        }
         notifService.deleteByPk(id)
         .then(notif => {
-            return notif;
+            res.status(200).json({
+                msg: "success delete data",
+                status: 200,
+            });
         })
         .catch(err => {
-            return err;
+            res.status(500).json({
+                msg: "error delete data",
+                status: 500,
+                err
+            });
         });
     },
-    getNotifByid(req,res) {
+    async getNotifByid(req,res) {
         const {id} = req.params
+        let check = await notifService.findByPk(id)
+        if(!check){
+            res.status(404).json({
+                msg: "data not found",
+                status: 404,
+            });
+            return
+        }
         notifService.findByPk(id)
         .then(notif => {
-            return notif;
+            res.status(200).json({
+                msg: "success get by id",
+                status: 200,
+                data: notif
+            });
         })
         .catch(err => {
-            return err;
+            res.status(500).json({
+                msg: "error get by id",
+                status: 500,
+                err
+            });
         });
     },
     async updateNotif(req,res) {
@@ -83,6 +123,30 @@ module.exports = {
             return insert
         } catch(err){
             return err
+        }
+    },
+    async createAdmin(req,res){
+        const {usersId,message} = req.body
+        const datas = {
+            id: uuid(),
+            usersId: usersId,
+            message: message,
+            isRead: false
+        }
+        const create = await notifService.createNotif(datas)
+        if(create){
+            const find = await notifService.findByUser(usersId)
+            global.io.sockets.in(usersId).emit('notify-update', find);
+            res.status(200).json({
+                msg: "notif updated",
+                status: 200,
+                data: create
+            });
+        }else{
+            res.status(500).json({
+                msg: "notif fails updated",
+                status: 500,
+            });
         }
     }
 }
